feat(6645): extract auth scope service api endpoint tests
