+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests.rs
@@ -36,6 +36,8 @@ use tokio::io::{AsyncReadExt, AsyncWriteExt};
 
 #[path = "service_api_endpoint_tests/balance_contract_tests.rs"]
 mod balance_contract_tests;
+#[path = "service_api_endpoint_tests/auth_scope_contract_tests.rs"]
+mod auth_scope_contract_tests;
 #[path = "service_api_endpoint_tests/websocket_contract_tests.rs"]
 mod websocket_contract_tests;
 
@@ -3089,668 +3091,6 @@ fn unit_service_api_endpoint_metrics_use_runtime_observability_when_present() {
     assert!(metrics_response.body.contains(&format!(
         "kamn_service_api_lifecycle_rejection_reason_code_count {expected_lifecycle_rejection_reason_code_count}"
     )));
-}
-
-#[test]
-fn integration_service_api_endpoint_accepts_case_variant_self_certifying_sender_did_binding() {
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34075".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 1,
-        idle_timeout_ms: 2_000,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let signer_public_key_hex =
-        service_auth_public_key_hex_from_private_key_hex(TEST_SERVICE_API_AUTH_PRIVATE_KEY_HEX)
-            .expect("service-auth public key should derive");
-    let sender_did = format!("kamn:did:agent:pkh-{signer_public_key_hex}");
-    let message_body = "{\"message\":\"hello\"}";
-    let state_hash = format!(
-        "service-api:{}:{}",
-        snapshot.chain_id.as_str(),
-        snapshot.chain_version.as_str()
-    );
-    let signature = service_api_request_signature_for_fields(
-        sender_did.as_str(),
-        1,
-        state_hash.as_str(),
-        message_body,
-    );
-    let response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", sender_did.as_str()),
-            ("X-KAMN-Request-Nonce", "1"),
-            ("X-KAMN-Request-Signature", signature.as_str()),
-            (
-                "x-kamn-signer-public-key",
-                signer_public_key_hex.to_uppercase().as_str(),
-            ),
-            ("X-KAMN-Authz-Scope", "messages:write"),
-        ],
-    );
-
-    assert!(response.contains("HTTP/1.1 202 Accepted"));
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after case-variant sender DID auth flow"
-    );
-}
-
-#[test]
-fn regression_service_api_endpoint_rejects_legacy_sender_binding_without_signer_public_key_header()
-{
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34076".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-    let state_hash = format!(
-        "service-api:{}:{}",
-        snapshot.chain_id.as_str(),
-        snapshot.chain_version.as_str()
-    );
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 1,
-        idle_timeout_ms: 2_000,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let sender_did = "kamn:did:agent:legacy-auth-binding";
-    let message_body = r#"{"recipient_did":"kamn:did:agent:legacy-auth-target","message":"hello"}"#;
-    let signature = service_auth_sign_with_private_key_hex(
-        sender_did,
-        1,
-        state_hash.as_str(),
-        message_body,
-        TEST_SERVICE_API_AUTH_PRIVATE_KEY_HEX,
-    )
-    .expect("raw legacy sender signature should derive");
-    let response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", sender_did),
-            ("X-KAMN-Request-Nonce", "1"),
-            ("X-KAMN-Request-Signature", signature.as_str()),
-        ],
-    );
-
-    assert!(response.contains("HTTP/1.1 401 Unauthorized"));
-    let error_payload = parse_error_envelope_from_http_response(response.as_str());
-    assert_eq!(
-        error_payload.reason_code,
-        "service_api_auth_signature_verification_failed"
-    );
-    assert!(
-        error_payload.message.contains("x-kamn-signer-public-key"),
-        "missing signer header rejection should explain the explicit signer binding contract"
-    );
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after rejecting legacy auth fallback"
-    );
-}
-
-#[test]
-fn unit_service_api_route_authz_matrix_matches_protected_and_public_paths() {
-    assert_eq!(
-        SERVICE_API_AUTH_REASON_TAXONOMY_VERSION,
-        "kamn.runtime.service-api-auth-reason-taxonomy.v1"
-    );
-    assert!(SERVICE_API_AUTH_REASON_CODES_CSV.contains(SERVICE_API_AUTH_MISSING_HEADER_REASON_CODE));
-    for row in service_api_route_authz_matrix_rows() {
-        assert_eq!(
-            crate::service_api_endpoint::route_requires_auth(row.method, row.path),
-            row.requires_auth,
-            "route authz matrix drift for {} {}",
-            row.method,
-            row.path
-        );
-    }
-}
-
-#[test]
-fn integration_service_api_endpoint_route_authz_matrix_rejects_protected_paths_without_headers() {
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    struct ObservedRouteOutcome {
-        method: String,
-        path: String,
-        status_line: String,
-        reason_code: Option<String>,
-    }
-
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34074".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let matrix_rows = service_api_route_authz_matrix_rows();
-    let rounds = 2_u64;
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: rounds * matrix_rows.len() as u64,
-        idle_timeout_ms: 2_500,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let mut baseline_outcomes: Option<Vec<ObservedRouteOutcome>> = None;
-    for _round in 0..rounds {
-        let mut outcomes = Vec::with_capacity(matrix_rows.len());
-        for row in &matrix_rows {
-            let response = send_http_request(bind_addr.as_str(), row.method, row.path, row.body);
-            assert!(
-                response.contains(row.expected_status_without_auth),
-                "unexpected authz matrix status for {} {}: expected {}, response={response}",
-                row.method,
-                row.path,
-                row.expected_status_without_auth
-            );
-            let reason_code = if row.requires_auth {
-                let payload = parse_error_envelope_from_http_response(response.as_str());
-                assert_eq!(payload.error, "unauthorized");
-                assert_eq!(
-                    payload.reason_code,
-                    SERVICE_API_AUTH_MISSING_HEADER_REASON_CODE
-                );
-                Some(payload.reason_code)
-            } else {
-                None
-            };
-            outcomes.push(ObservedRouteOutcome {
-                method: row.method.to_owned(),
-                path: row.path.to_owned(),
-                status_line: row.expected_status_without_auth.to_owned(),
-                reason_code,
-            });
-        }
-        if let Some(baseline) = baseline_outcomes.as_ref() {
-            assert_eq!(
-                outcomes, *baseline,
-                "route authz outcomes must remain deterministic across rounds"
-            );
-        } else {
-            baseline_outcomes = Some(outcomes);
-        }
-    }
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after route authz matrix validation"
-    );
-}
-
-#[test]
-fn unit_service_api_scope_policy_fixture_parser_contract() {
-    let (metadata, rows) = parse_service_api_scope_policy_fixture(SERVICE_API_SCOPE_POLICY_FIXTURE);
-    assert_eq!(
-        metadata
-            .get("scope_policy_fixture_matrix_schema_version")
-            .map(String::as_str),
-        Some(SERVICE_API_SCOPE_POLICY_FIXTURE_SCHEMA_VERSION)
-    );
-    assert_eq!(
-        metadata
-            .get("scope_policy_reason_taxonomy_version")
-            .map(String::as_str),
-        Some(SERVICE_API_SCOPE_POLICY_REASON_TAXONOMY_VERSION)
-    );
-    assert_eq!(
-        metadata
-            .get("scope_policy_reason_codes_csv")
-            .map(String::as_str),
-        Some(SERVICE_API_SCOPE_POLICY_REASON_CODES_CSV)
-    );
-    assert!(
-        rows.len() >= 6,
-        "scope policy fixture matrix should provide representative coverage"
-    );
-    assert!(
-        rows.iter().any(|row| {
-            row.method == "POST"
-                && row.path == "/v1/messages/send"
-                && row.scope == "messages:write"
-                && row.expected == "allow"
-        }),
-        "fixture should include allow case for message send write scope"
-    );
-    assert!(
-        rows.iter().any(|row| {
-            row.method == "POST"
-                && row.path == "/v1/messages/send"
-                && row.scope == "messages:read"
-                && row.expected == "deny"
-        }),
-        "fixture should include deny case for message send read scope"
-    );
-}
-
-#[test]
-fn functional_service_api_scope_policy_fixture_rows_match_route_scope_mapping() {
-    let (_, rows) = parse_service_api_scope_policy_fixture(SERVICE_API_SCOPE_POLICY_FIXTURE);
-    for row in rows {
-        let expected_scope = required_scope_for_test_route(row.method.as_str(), row.path.as_str())
-            .expect("fixture rows should target protected routes only");
-        if row.expected == "allow" {
-            assert_eq!(
-                row.scope, expected_scope,
-                "allow fixture row scope must match required route scope"
-            );
-        } else if row.expected == "deny" {
-            assert_ne!(
-                row.scope, expected_scope,
-                "deny fixture row scope must not match required route scope"
-            );
-        } else {
-            panic!("scope fixture expected field must be allow|deny");
-        }
-    }
-}
-
-#[test]
-fn integration_service_api_endpoint_scope_policy_rejects_missing_invalid_and_mismatched_scopes() {
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34075".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 4,
-        idle_timeout_ms: 2_500,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let message_body = "{\"message\":\"scope-policy-check\"}";
-    let sender_did = "kamn:did:agent:test-client-scope-policy";
-    let bound_sender_did = test_service_api_sender_did(sender_did);
-    let signer_public_key_hex = test_service_api_auth_public_key_hex();
-    let state_hash = format!(
-        "service-api:{}:{}",
-        snapshot.chain_id.as_str(),
-        snapshot.chain_version.as_str()
-    );
-    let signature_missing_scope = service_api_request_signature_for_fields(
-        sender_did,
-        9101,
-        state_hash.as_str(),
-        message_body,
-    );
-    let signature_invalid_scope = service_api_request_signature_for_fields(
-        sender_did,
-        9102,
-        state_hash.as_str(),
-        message_body,
-    );
-    let signature_mismatch_scope = service_api_request_signature_for_fields(
-        sender_did,
-        9103,
-        state_hash.as_str(),
-        message_body,
-    );
-    let signature_allowed_scope = service_api_request_signature_for_fields(
-        sender_did,
-        9104,
-        state_hash.as_str(),
-        message_body,
-    );
-
-    let missing_scope_response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", bound_sender_did.as_str()),
-            ("X-KAMN-Request-Nonce", "9101"),
-            ("X-KAMN-Request-Signature", signature_missing_scope.as_str()),
-            ("X-KAMN-Signer-Public-Key", signer_public_key_hex.as_str()),
-        ],
-    );
-    assert!(missing_scope_response.contains("HTTP/1.1 401 Unauthorized"));
-    let missing_scope_payload =
-        parse_error_envelope_from_http_response(missing_scope_response.as_str());
-    assert_eq!(missing_scope_payload.error, "unauthorized");
-    assert_eq!(
-        missing_scope_payload.reason_code,
-        SERVICE_API_AUTH_SCOPE_HEADER_MISSING_REASON_CODE
-    );
-
-    let invalid_scope_response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", bound_sender_did.as_str()),
-            ("X-KAMN-Request-Nonce", "9102"),
-            ("X-KAMN-Request-Signature", signature_invalid_scope.as_str()),
-            ("X-KAMN-Signer-Public-Key", signer_public_key_hex.as_str()),
-            ("X-KAMN-Authz-Scope", ""),
-        ],
-    );
-    assert!(invalid_scope_response.contains("HTTP/1.1 401 Unauthorized"));
-    let invalid_scope_payload =
-        parse_error_envelope_from_http_response(invalid_scope_response.as_str());
-    assert_eq!(invalid_scope_payload.error, "unauthorized");
-    assert_eq!(
-        invalid_scope_payload.reason_code,
-        SERVICE_API_AUTH_SCOPE_INVALID_REASON_CODE
-    );
-
-    let mismatch_scope_response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", bound_sender_did.as_str()),
-            ("X-KAMN-Request-Nonce", "9103"),
-            (
-                "X-KAMN-Request-Signature",
-                signature_mismatch_scope.as_str(),
-            ),
-            ("X-KAMN-Signer-Public-Key", signer_public_key_hex.as_str()),
-            ("X-KAMN-Authz-Scope", "messages:read"),
-        ],
-    );
-    assert!(mismatch_scope_response.contains("HTTP/1.1 401 Unauthorized"));
-    let mismatch_scope_payload =
-        parse_error_envelope_from_http_response(mismatch_scope_response.as_str());
-    assert_eq!(mismatch_scope_payload.error, "unauthorized");
-    assert_eq!(
-        mismatch_scope_payload.reason_code,
-        SERVICE_API_AUTH_SCOPE_ROUTE_MISMATCH_REASON_CODE
-    );
-
-    let allowed_scope_response = send_http_request_with_headers_raw(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        message_body,
-        &[
-            ("X-KAMN-Sender-DID", bound_sender_did.as_str()),
-            ("X-KAMN-Request-Nonce", "9104"),
-            ("X-KAMN-Request-Signature", signature_allowed_scope.as_str()),
-            ("X-KAMN-Signer-Public-Key", signer_public_key_hex.as_str()),
-            ("X-KAMN-Authz-Scope", "messages:write"),
-        ],
-    );
-    assert!(allowed_scope_response.contains("HTTP/1.1 202 Accepted"));
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after scope policy checks"
-    );
-}
-
-#[test]
-fn integration_service_api_endpoint_rejects_missing_request_auth_headers() {
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34053".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 1,
-        idle_timeout_ms: 2_000,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let unauth_response = send_http_request(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        "{\"message\":\"hello\"}",
-    );
-    assert!(unauth_response.contains("HTTP/1.1 401 Unauthorized"));
-    let unauth_payload = parse_error_envelope_from_http_response(unauth_response.as_str());
-    assert_eq!(unauth_payload.error, "unauthorized");
-    assert_eq!(
-        unauth_payload.reason_code,
-        "service_api_auth_sender_did_header_missing"
-    );
-    assert!(unauth_payload.message.contains("x-kamn-sender-did"));
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after configured request budget"
-    );
-}
-
-#[test]
-fn integration_service_api_endpoint_rejects_legacy_deterministic_signature_profile() {
-    let _env = acquire_service_api_test_env();
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34079".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 1,
-        idle_timeout_ms: 2_000,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let sender_did = "kamn:did:agent:test-client-legacy-signature";
-    let nonce = 1_u64;
-    let payload = r#"{"message":"legacy-signature"}"#;
-    let state_hash = format!(
-        "service-api:{}:{}",
-        snapshot.chain_id.as_str(),
-        snapshot.chain_version.as_str()
-    );
-    let legacy_signature =
-        kamn_core::legacy_signature_for_fields(sender_did, nonce, state_hash.as_str(), payload);
-
-    let response = send_http_request_with_headers(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        payload,
-        &[
-            ("X-KAMN-Sender-DID", sender_did),
-            ("X-KAMN-Request-Nonce", "1"),
-            ("X-KAMN-Request-Signature", legacy_signature.as_str()),
-        ],
-    );
-    assert!(response.contains("HTTP/1.1 401 Unauthorized"));
-    let error_payload = parse_error_envelope_from_http_response(response.as_str());
-    assert_eq!(
-        error_payload.reason_code,
-        "service_api_auth_signature_verification_failed"
-    );
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after configured request budget"
-    );
-}
-
-#[test]
-fn regression_service_api_endpoint_rejects_legacy_signature_when_toggle_env_is_true() {
-    let _env = acquire_service_api_test_env();
-    let _legacy_toggle_guard = EnvVarGuard::set(
-        "KAMN_SERVICE_API_AUTH_ALLOW_LEGACY_SIGNATURES",
-        Some("true"),
-    );
-    let parsed = parse_args(vec![
-        "kamn-node".to_owned(),
-        "--role".to_owned(),
-        "processor".to_owned(),
-        "--runtime-mode".to_owned(),
-        "api".to_owned(),
-        "--api-bind".to_owned(),
-        "127.0.0.1:34095".to_owned(),
-    ])
-    .expect("api args should parse");
-    let report = execute(parsed).expect("api execution should succeed");
-    let snapshot = build_service_api_snapshot(&report);
-
-    let bind_addr = reserve_loopback_addr();
-    let endpoint_config = ServiceApiEndpointConfig {
-        bind_addr: bind_addr.clone(),
-        max_requests: 1,
-        idle_timeout_ms: 2_000,
-        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
-        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
-        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
-    };
-    let server_snapshot = snapshot.clone();
-    let server =
-        thread::spawn(move || serve_service_api_endpoint(&endpoint_config, &server_snapshot));
-    wait_for_endpoint_ready(bind_addr.as_str());
-
-    let sender_did = "kamn:did:agent:test-client-legacy-toggle-true";
-    let nonce = 1_u64;
-    let payload = r#"{"message":"legacy-signature-toggle-true"}"#;
-    let state_hash = format!(
-        "service-api:{}:{}",
-        snapshot.chain_id.as_str(),
-        snapshot.chain_version.as_str()
-    );
-    let legacy_signature =
-        kamn_core::legacy_signature_for_fields(sender_did, nonce, state_hash.as_str(), payload);
-
-    let response = send_http_request_with_headers(
-        bind_addr.as_str(),
-        "POST",
-        "/v1/messages/send",
-        payload,
-        &[
-            ("X-KAMN-Sender-DID", sender_did),
-            ("X-KAMN-Request-Nonce", "1"),
-            ("X-KAMN-Request-Signature", legacy_signature.as_str()),
-        ],
-    );
-    assert!(response.contains("HTTP/1.1 401 Unauthorized"));
-    let error_payload = parse_error_envelope_from_http_response(response.as_str());
-    assert_eq!(
-        error_payload.reason_code,
-        "service_api_auth_signature_verification_failed"
-    );
-
-    let server_result = server.join().expect("endpoint thread should complete");
-    assert!(
-        server_result.is_ok(),
-        "service api endpoint should stop cleanly after configured request budget"
-    );
 }
 
 #[test]
