+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/auth_scope_contract_tests.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+use support::{join_service_api_server, service_api_request_state_hash, start_service_api_server};
+
+#[path = "auth_scope_contract_tests/support.rs"]
+mod support;
+#[path = "auth_scope_contract_tests/auth_binding_contract_tests.rs"]
+mod auth_binding_contract_tests;
+#[path = "auth_scope_contract_tests/route_scope_policy_contract_tests.rs"]
+mod route_scope_policy_contract_tests;
+#[path = "auth_scope_contract_tests/legacy_signature_contract_tests.rs"]
+mod legacy_signature_contract_tests;
