+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/auth_scope_contract_tests/support.rs
@@ -1,0 +1,57 @@
+use super::*;
+
+pub(super) fn start_service_api_server(
+    api_bind: &str,
+    max_requests: u64,
+) -> (
+    crate::service_api_endpoint::ServiceApiSnapshot,
+    String,
+    thread::JoinHandle<Result<(), String>>,
+) {
+    let _env = acquire_service_api_test_env();
+    let parsed = parse_args(vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "api".to_owned(),
+        "--api-bind".to_owned(),
+        api_bind.to_owned(),
+    ])
+    .expect("api args should parse");
+    let report = execute(parsed).expect("api execution should succeed");
+    let snapshot = build_service_api_snapshot(&report);
+    let bind_addr = reserve_loopback_addr();
+    let endpoint_config = ServiceApiEndpointConfig {
+        bind_addr: bind_addr.clone(),
+        max_requests,
+        idle_timeout_ms: 2_500,
+        body_limit_bytes: DEFAULT_SERVICE_API_BODY_LIMIT_BYTES,
+        concurrency_limit: DEFAULT_SERVICE_API_CONCURRENCY_LIMIT,
+        rate_limit_per_second: DEFAULT_SERVICE_API_RATE_LIMIT_PER_SECOND,
+    };
+    let server_snapshot = snapshot.clone();
+    let server = thread::spawn(move || {
+        serve_service_api_endpoint(&endpoint_config, &server_snapshot)
+    });
+    wait_for_endpoint_ready(bind_addr.as_str());
+    (snapshot, bind_addr, server)
+}
+
+pub(super) fn join_service_api_server(
+    server: thread::JoinHandle<Result<(), String>>,
+    message: &str,
+) {
+    let server_result = server.join().expect("endpoint thread should complete");
+    assert!(server_result.is_ok(), "{message}");
+}
+
+pub(super) fn service_api_request_state_hash(
+    snapshot: &crate::service_api_endpoint::ServiceApiSnapshot,
+) -> String {
+    format!(
+        "service-api:{}:{}",
+        snapshot.chain_id.as_str(),
+        snapshot.chain_version.as_str()
+    )
+}
